pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "odcv-fus-dpo"
version = "0.1.0"
description = "Implementation scaffold for first-unsafe-step counterfactual DPO on ODCV-Bench."
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
  "pydantic>=2.6",
  "orjson>=3.10",
  "PyYAML>=6.0",
  "jsonschema>=4.22",
  "typer>=0.12",
  "rich>=13.7",
  "pandas>=2.2",
  "scipy>=1.13",
  "matplotlib>=3.8",
]

[project.optional-dependencies]
train = [
  "transformers>=4.49",
  "datasets>=3.2",
  "accelerate>=1.3",
  "peft>=0.14",
  "trl>=0.15",
  "bitsandbytes>=0.45",
  "vllm>=0.7",
]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
  "mypy>=1.10",
]

[tool.setuptools.packages.find]
where = ["src"]
